change precision for decimals; add foreign_index to Orders.customer

+++ db/migrate/20190113001100_create_books.rb
@@ -4,7 +4,7 @@ class CreateBooks < ActiveRecord::Migration[5.2]
       t.string :title, null: false
       t.integer :year_published
       t.string :isbn
-      t.decimal :price, precision: 16, scale: 2, default: 0, null: false
+      t.decimal :price, precision: 10, scale: 2, default: 0, null: false
 
       t.boolean :out_of_print, default: false
